+++ various_calculations/averages.py
