+++ various_calculations/get_rest_groups.py
