+++ various_calculations/poss_groups.py
